task-8-deployment/Streamlit/requirements.txt
beautifulsoup4==4.9.3
fastai==2.4
mediacloud-cliff==2.6.1
newspaper3k==0.2.8
pandas==1.2.4
pygooglenews==0.1.2
requests==2.26.0
spacy==2.2.4
streamlit==0.85.1
torch==1.7.1
keras==2.3.1
tensorflow==2.4.1
numpy==1.19.4
# the following is for fixing Twint, see https://github.com/twintproject/twint/issues/915
git+https://github.com/twintproject/twint.git@origin/master#egg=twint
pickle5==0.0.11
nltk==3.6.2
